requirements.txt
flask>=3.0.0
openai==0.28.1
python-dotenv>=1.0.0
werkzeug>=3.0.0
numpy>=1.26.0
requests>=2.31.0
psutil>=5.9.0
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.4
gunicorn>=21.2.0
